Extract the First Data Release (Vintage)

Filters the input dataset to return the earliest release (or vintage) for each time period.

Usage

get_first_release(df)

Arguments

df

A data frame containing data vintages. The data frame must include the columns pub_date (publication date of the release) and time (the corresponding time period for the data).

Value

A filtered data frame containing only the first release(s). The resulting data frame is assigned the class tbl_nth_rel to indicate its structure.

Details

For each time period, the function identifies the release with the earliest publication date (pub_date). A new column release is added and labels all rows in the resulting data frame as release_0.

Examples

# Example data
df <- reviser::gdp

# Get the first release for each time period
first_release <- get_first_release(df)