Create Matrices for the Kishor-Koenig (KK) or Related Models

Constructs the matrices \(I\), \(F\), \(G\), \(R\), and \(H\) used in state-space models, specifically for the Kishor-Koenig (KK), Howrey, or Classical frameworks.

Usage

kk_matrices(e, model, params = NULL, type = "numeric")

Arguments

e

Integer. The number of efficiency gaps (lags) in the model. Must be greater than 0.

model

Character. Specifies the type of model to use. Options are:

"Kishor-Koenig" or "KK"

Uses the Kishor-Koenig framework with \(e \times (e+1)\) parameters for the \(G\) matrix.

"Howrey"

Uses the Howrey framework with \(e \times e\) parameters for the \(G\) matrix.

"Classical"

Uses a diagonal identity matrix for \(G\).

params

Numeric vector (optional). A vector of parameters to initialize the matrices. If NULL, default values are used:

type = "numeric"

A vector of params must be supplied.

type = "character"

Initializes named parameters as NA_real_.

If provided, the length of params must match the number of parameters required by the specified model.

type

Character. Specifies the type of matrices returned. Options are:

"numeric"

Returns numeric matrices with parameter values.

"character"

Returns character matrices with parameter names. If params is provided, it is ignored.

Value

A list containing the following components:

II

Identity matrix (\(I\)). Size: \((e+1) \times (e+1)\).

FF

State transition matrix (\(F\)). Size: \((e+1) \times (e+1)\).

GG

Control matrix (\(G\)). Size depends on the model and e.

R

State noise covariance matrix (\(R\)). Size: \((e+1) \times (e+1)\).

H

Observation noise covariance matrix (\(H\)). Size: \((e+1) \times (e+1)\).

params

The vector of parameters used to construct the matrices, including their names.

Examples

# Example 1: Kishor-Koenig model with character matrices
matrices <- kk_matrices(e = 3, model = "KK", type = "character")
str(matrices)
#> List of 6
#>  $ II    : chr [1:4, 1:4] "1" "0" "0" "0" ...
#>  $ FF    : chr [1:4, 1:4] "0" "0" "0" "0" ...
#>  $ GG    : chr [1:4, 1:4] "1" "G2_3" "G1_3" "G0_3" ...
#>  $ R     : chr [1:4, 1:4] "0" "0" "0" "0" ...
#>  $ H     : chr [1:4, 1:4] "0" "0" "0" "0" ...
#>  $ params: Named num [1:17] NA NA NA NA NA NA NA NA NA NA ...
#>   ..- attr(*, "names")= chr [1:17] "F0" "G0_0" "G0_1" "G0_2" ...

# Example 2: Kishor-Koenig model with e = 2
params <- rep(0.1, 17)
names(params) <- names(matrices$params)
matrices <- kk_matrices(e = 3, params = params,  model = "KK", type = "numeric")
str(matrices)
#> List of 6
#>  $ II    : num [1:4, 1:4] 1 0 0 0 0 1 0 0 0 0 ...
#>  $ FF    : num [1:4, 1:4] 0 0 0 0 1 0 0 0 0 1 ...
#>  $ GG    : num [1:4, 1:4] 1 0.1 0.1 0.1 0 0.1 0.1 0.1 0 0.1 ...
#>  $ R     : num [1:4, 1:4] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ H     : num [1:4, 1:4] 0 0 0 0 0 0.1 0 0 0 0 ...
#>  $ params: Named num [1:17] 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 0.1 ...
#>   ..- attr(*, "names")= chr [1:17] "F0" "G0_0" "G0_1" "G0_2" ...