Rename Columns to Align with Package Standards

Renames columns in a data frame or tibble to align with the conventions used in this package. Converts the renamed columns to the appropriate data types.

Usage

vintages_rename(
  df,
  col_time = NULL,
  col_pub_date = NULL,
  col_value = NULL,
  col_release = NULL,
  col_id = NULL
)

Arguments

df

A data frame or tibble containing the data to be renamed.

col_time

Optional. The name of the column to be renamed as time. The time column represents observation dates and will be converted to Date format.

col_pub_date

Optional. The name of the column to be renamed as pub_date. The pub_date column represents release dates and will be converted to Date format.

col_value

Optional. The name of the column to be renamed as value. The value column represents the observed values and will be converted to numeric.

col_release

Optional. The name of the column to be renamed as release. The id column is used as an identifier and will be converted to character format.

col_id

Optional. The name of the column to be renamed as id.

Value

A data frame or tibble with the renamed columns and their respective data types converted (if specified). The original class of the input object is preserved.

Details

The function checks the validity of the input data frame and ensures that at least one column is specified for renaming. If a column is renamed, it is also converted to the expected data type:

time and pub_date are converted to Date.

value is converted to numeric.

release and id are converted to character.

Examples

# Example data
data <- tibble::tibble(
  observation_date = seq.Date(as.Date("2020-01-01"), as.Date("2020-06-01"), by = "month"),
  release_date = seq.Date(as.Date("2020-01-15"), as.Date("2020-06-15"), by = "month"),
  observed_value = rnorm(6),
  identifier = rep("A", 6)
)

# Rename columns
renamed_data <- vintages_rename(
  data,
  col_time = observation_date,
  col_pub_date = release_date,
  col_value = observed_value,
  col_id = identifier
)