Generalized Kishor-Koenig Model for Nowcasting

Implements a generalized Kishor-Koenig (KK) model for nowcasting and forecasting with state-space models, allowing for multiple vintages of data, efficient estimation, and Kalman filtering and smoothing.

Usage

kk_nowcast(df, e, h = 0, model = "Kishor-Koenig", trace = 0)

Arguments

df

A data frame containing the time series data in either "long" or "wide" format. It must include columns for the time index and the different release vintages.

e

An integer indicating the number of data vintages to include in the model. Must be greater than 0.

h

An integer specifying the forecast horizon. Default is 0, which implies no forecasts. Must be greater than or equal to 0.

model

A string specifying the type of model to use. Options are: - "Kishor-Koenig" or "KK" (default): Full Kishor-Koenig model. - "Howrey": Howrey's simplified framework. - "Classical": Classical model without vintage effects.

trace

An integer controlling the level of output for the optimization procedure. Default is 0 (minimal output).

Value

A list with the following components:

forecast_states

A tibble of forecasted state variables, including the forecast dates. Returned only if h > 0.

filtered_states

A tibble of filtered state variables based on the Kalman filter.

observations

A tibble of the observed variables used in the model.

forecast_observation

A tibble of forecasted observations. Returned only if h > 0.

smoothed_states

A tibble of smoothed state variables based on the Kalman smoother.

kk_model_mat

A list of KK model matrices (e.g., transition and observation matrices).

ss_model_mat

A list of state-space model matrices derived from the KK model.

params

Estimated model parameters, including covariance terms.

fit

The fitted model object from the SUR estimation procedure.

Details

The function supports multiple models, including the full Kishor-Koenig framework, Howrey's model, and a classical approach. It handles data preprocessing, estimation of system equations using Seemingly Unrelated Regressions (SUR), and application of the Kalman filter and smoother.

The function requires well-structured input data with multiple vintages. The time series must be regular, and the function automatically checks and transforms the data if needed.

Examples

# Example usage:
df <- get_nth_release(tsbox::ts_span(tsbox::ts_pc(reviser::gdp_us), start = "1980-01-01"), n = 0:1)
e <- 1  # Number of efficient release
h <- 2  # Forecast horizon
model_result <- kk_nowcast(df, e, h = h, model = "Kishor-Koenig")
#> Warning: Ignoring columns: pub_date

model_result$params
#>           F0          G01          G00           v0         eps0 
#>  0.161673050 -0.007862814  0.974344214  1.542794735  0.035485021