Summary of Efficient Release Models

Provides a detailed summary of the regression model and hypothesis test for the first efficient release identified by the get_first_efficient_release function.

Usage

# S3 method for class 'lst_efficient'
summary(object, ...)

Arguments

object

An output object from the get_first_efficient_release function. The object must be of class list_eff_rel.

...

Additional arguments (not used).

Value

This function does not return a value. It prints the summary to the console.

Details

This function prints the following information:

The index of the first efficient release.

A summary of the regression model fitted for the efficient release, which includes coefficients, R-squared values, and other relevant statistics.

The hypothesis test results for the efficient release, showing the test statistic and p-value for the null hypothesis of unbiasedness and efficiency.

The function assumes the object includes:

e: The index of the first efficient release (0-based).

models: A list of linear regression models for each release.

tests: A list of hypothesis test results corresponding to each release.

Examples

# Example usage
df <- get_nth_release(tsbox::ts_pc(reviser::gdp), n = 1:4)

final_release <- get_nth_release(tsbox::ts_pc(reviser::gdp), n = 10)

# Identify the first efficient release
result <- get_first_efficient_release(df, final_release, significance = 0.05)
summary(result)
#> Efficient release:  0 
#> 
#> Model summary: 
#> 
#> Call:
#> stats::lm(formula = formula, data = df_wide)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.90728 -0.20657  0.00298  0.19586  0.86728 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  0.01646    0.02401   0.686    0.494    
#> release_1    0.96538    0.01862  51.836   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 0.3026 on 206 degrees of freedom
#>   (9 observations deleted due to missingness)
#> Multiple R-squared:  0.9288,	Adjusted R-squared:  0.9284 
#> F-statistic:  2687 on 1 and 206 DF,  p-value: < 2.2e-16
#> 
#> 
#> Test summary: 
#> 
#> Linear hypothesis test:
#> (Intercept) = 0
#> release_1 = 1
#> 
#> Model 1: restricted model
#> Model 2: final ~ release_1
#> 
#> Note: Coefficient covariance matrix supplied.
#> 
#>   Res.Df Df      F Pr(>F)
#> 1    208                 
#> 2    206  2 1.8886 0.1539